(PC-28359): Fix tuile page offre
